Move assets folder to project root
Currently the "assets" folder is located in the "src" folder. It has
been moved to the project root to clarify that the files are not source
parts of the color palette.

This change will break many links to the contained asset files in the
port project repositories. Each port project is responsible to adapt to
the updated assets paths.

GH-83