Specify distributed npm package files
Currently the whole project repository is published to the npm registry
which results in a unnecessary large package size. The "files" field
(array) (1) should be added to the "package.json" to optimize the
distributed package content and decrease the size drastically.

The native color swatches files (2) should not be distributed.

References:

  (1) https://docs.npmjs.com/misc/developers#keeping-files-out-of-your-package
  (2) https://arcticicestudio.github.io/nord/introduction/color-swatches.html

GH-84