`nordtheme` organization migration
As part of the "Northern Post — The state and roadmap of Nord" [2]
announcement, this repository will be migrated to the `nordtheme`
GitHub organization [1].
This commit tracks the actual move as well as preparations steps to do
so. The detailed plan, tasklists that serve as "epics" [3], will follow
later on for all Nord repositories, published and announced on all "new"
and current community platforms.

[1]: https://github.com/arcticicestudio/nord/issues/180
[2]: https://github.com/nordtheme
[3]: https://docs.github.com/en/get-started/writing-on-github/working-with-advanced-formatting/about-task-lists

GH-181